refactor: 자녀 선택에 따른 hasAlreadyBeenFetched 로직 함수로 분리
